getToppings method was improved: only one query is used

+++ src/main/kotlin/com/github/mrf1n/pizzeria/controller/PizzeriaController.kt
@@ -24,7 +24,7 @@ class PizzeriaController(
     }
 
     @GetMapping
-    fun getToppings(): Map<String, Int> = toppingsService.getToppings()
+    fun getToppings(): Map<String, Long> = toppingsService.getToppings()
 
     @GetMapping("/{email}")
     fun getToppingsByEmail(@Valid @Email @PathVariable("email") email: String): Set<String> =
