+++ src/main/kotlin/com/github/mrf1n/pizzeria/service/ToppingsService.kt
@@ -4,7 +4,7 @@ interface ToppingsService {
 
     fun addToppings(customerEmail: String, toppings: Set<String>)
 
-    fun getToppings(): Map<String, Int>
+    fun getToppings(): Map<String, Long>
 
     fun getToppingsForCustomer(email: String): Set<String>
 }
